Move log and audit workers out of eventbus
