+++ internal/auditworker/worker.go
@@ -1,28 +1,16 @@
-package eventbus
+package auditworker
 
 import (
 	"context"
 	"log"
 
 	dbpkg "github.com/arran4/goa4web/internal/db"
+	"github.com/arran4/goa4web/internal/eventbus"
 	"github.com/arran4/goa4web/internal/tasks"
 )
 
-// LogWorker listens on the bus and logs all received events.
-func LogWorker(ctx context.Context, bus *Bus) {
-	ch := bus.Subscribe()
-	for {
-		select {
-		case evt := <-ch:
-			log.Printf("event path=%s task=%s uid=%d data=%v", evt.Path, evt.Task, evt.UserID, evt.Data)
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-// AuditWorker records bus events into the audit_log table.
-func AuditWorker(ctx context.Context, bus *Bus, q *dbpkg.Queries) {
+// Worker records bus events into the audit_log table.
+func Worker(ctx context.Context, bus *eventbus.Bus, q *dbpkg.Queries) {
 	if q == nil || bus == nil {
 		return
 	}
